requirements.txt
streamlit>=1.25.0
torch>=1.12.0
torchvision>=0.13.0
pillow
numpy
pandas
